updatd StudentController

+++ src/main/java/com/rogers/demo/controller/StudentController.java
@@ -1,6 +1,7 @@
 package com.rogers.demo.controller;
 
 import com.rogers.demo.entity.Student;
+import com.rogers.demo.exception.StudentNotFoundException;
 import com.rogers.demo.exception.StudentNullException;
 import com.rogers.demo.service.StudentService;
 import io.swagger.annotations.ApiOperation;
@@ -42,7 +43,7 @@ public class StudentController {
     @PostMapping({"student/add"})
     public Student addStudent(@RequestBody Student student){
         if(null==student || null==student.getName()){
-            throw new StudentNullException();
+            throw new StudentNullException("Student can not be null");
         }
         studentService.addNewStudent(student);
         return student;
@@ -57,6 +58,10 @@ public class StudentController {
     })
     @DeleteMapping({"student/remove/{studentId:\\d+}"})
     public void deleteStudent(@PathVariable final Long studentId){
+        Student student=studentService.getStudentById(studentId);
+        if(student==null){
+            throw new StudentNotFoundException(studentId);
+        }
         studentService.deleteStudent(studentId);
     }
 }
